The old Pure API is now Web Services in this repo.
To help separate support for the old Pure read-only API from the
new read-write API, we follow Elsevier's lead and refer to the old
API as "Pure Web Services" and the new API as the "Pure API". The
code included here under experts/pure/web_services is the same as
that in the old pureapi repo, with names changed accordingly to
reflect the new naming scheme.